Schedule job per resource collection and discover resources for the collection name
